+++ app/registerFn.php
@@ -1,0 +1,3 @@
+<?php
+
+/** Here you can add your own global available functions */